Add ability to redact/modify parts of log event text
